+++ src_new/triggers/levi_ataxia/for_levi/leviticus/mnemosyne/001_Run_Start.lua
@@ -44,6 +44,7 @@ mnemRollHide = false  -- boons reset each run
 mnemBloodscent = false  -- boons reset each run
 mnemKaiUnleashed = false  -- boons reset each run
 mnemSenselessFlurry = false  -- boons reset each run
+mnemSpiritRend = false  -- boons reset each run
 psionPanoply = false  -- boons reset each run
 dragonMightSycaerunax = false  -- boons reset each run
 dragonRampage = false  -- boons reset each run